added and implemented dictionary repository

+++ shared/src/commonMain/kotlin/ru/dinarastepina/nivkh/data/local/LocalDataSource.kt
@@ -8,11 +8,11 @@ interface LocalDataSource {
 
     suspend fun searchNivkhTranslations(limit: Int, offset: Int, query: String): List<Long>
 
-    suspend fun searchNivkhWords(ids: List<Int>): List<NivkhWord>
+    suspend fun searchNivkhWords(ids: List<Long>): List<NivkhWord>
 
     suspend fun getAllRussianWords(limit: Int, offset: Int): List<RussianWord>
 
     suspend fun searchRussianTranslations(limit: Int, offset: Int, query: String): List<Long>
 
-    suspend fun searchRussianWords(ids: List<Int>): List<RussianWord>
+    suspend fun searchRussianWords(ids: List<Long>): List<RussianWord>
 }
